fix: 토큰 유틸 함수에 SSR 대응 안전 장치 추가
