--- src/api/tokenUtils.ts
@@ -1,14 +1,0 @@
-const ACCESS_KEY = 'access_token'
-const REFRESH_KEY = 'refresh_token'
-
-export const saveTokens = (access: string, refresh: string) => {
-  localStorage.setItem(ACCESS_KEY, access)
-  localStorage.setItem(REFRESH_KEY, refresh)
-}
-
-export const getAccessToken = () => localStorage.getItem(ACCESS_KEY)
-export const getRefreshToken = () => localStorage.getItem(REFRESH_KEY)
-export const clearTokens = () => {
-  localStorage.removeItem(ACCESS_KEY)
-  localStorage.removeItem(REFRESH_KEY)
-}
